refactor: move internal util to internal/fs
